Views: move LSPS1 dir
